Fully run and return from user mode function.
I can now allocate the stack, give the user permission to rw it and
enter the user function, allowing it to execute to completion.

Upon return, a svc call is used to raise permissions and then switch the
stack pointer back.  The current implementation is a complete hack and
needs to be fixed, but it worked in a jiffy.  As it stands, an user mode
function code request raised privileges and it would get them.