Moved .kernel section to fix MPU permission errors
MPU regions must be aligned to the size of the region, and the kernel
section was not, causing the kernel permissions to effect some of the
text section.  By putting the kernel section first, the text section can
be aligned afterwards to avoid any overflow from the kernel's MPU
permissions.