Tasks that return are successfully cleaned up
Their memory still needs to be garbage collected, by they are removed
from the task list and things go on without them.

Enabling of the FPU was also added to the C startup.

Macro definitions are now available in gdb thanks to -g3

k_currentTask is now the taskNode, making it easy to find the nearby
tasks.