Made task switching more robust against interruption
Certain operations, in particular initial tasking set up and the removal
of a completed task, were vulnerable to pre-emption, which in the case
of a systick, could cause them to crash.

In tasking set up, this has been avoided by not starting the systick
timer until just before task switching has begun.  It may be worth
looking into moving start_task_switching into an interrupt or disabling
interrupts, as it is still vulnerable to pre-emption.

Most of end_task has been moved to svc_handler, where it not only cannot
be interrupted by systick or pendsv, but it can be simpler, as some of
the restoration is taken care of by the exception.