Clean up lr saves and restores in context.c
Similar to what was suggested in pull request #3, but attempts to be
more fault tolerant be telling the compiler to set aside a register just
for storing the lr value we want.  While the C is ugly, it should be
optimization safe, and compiles to simple "mov r9, lr" and "mov lr, r9"
instructions.