Fix/workaround ipctest
Added waiting task node to semaphore struct to indicate the highest
priority task waiting for a semaphore.  This task is switched to on
release if it is >= priority of current task.  This prevents ipctest
from trying thousands of times, as shell reacquires its semaphores
almost immediately after releasing them.

In that way, it is a workaround.  There must have been some other edge
condition that actually caused the looping to fail, and this just avoids
all of the looping.

Unfortunately, I had to optimize(0) swap_task.  Setting an asm register
does not reserve that register, so my use of r0 clobbered node, and I
had to switch to r1.  The switch was trivial, but for some reason at
-O2, the function returns from the swap with a stack pointer off by 4.
The asm is almost identical between the r0 and r1 versions, and I am at
a loss as to why it makes a difference.