Fix buddy_merge fatal bug.
When adding merged buddies back to the list, buddy_merge always added
the node requested to be merged to the list, even though it may be the
upper half of the buddy.

This became evident in ipctest, which was allocated an improperly merged
address, and then clobbered someone else's memory.  ipctest now works
flawlessly, though it still leaks a small amount of memory.

Added some NULL checking for resources and initialize the resource array
to NULL.

Since ipctest spawns memreader, it never closes the shared memory, and
shell may run out of resource descriptors.