Workaround for periodic printing bug
getc() will release its semaphore before yielding and reacquire it after
restoring.  This allows others to come in and use the semaphore, and
they may steal the character getc() was waiting for.  However, this
prevents near deadloops as getc() waits forever for a character.

Unfortunately, this has not fixed the fact that near deaploops do
eventually crash, which could be a major problem.  This just avoids the
deadloops.

In acquire, the 0: label was moved above the inline assembly block to
ensure that addr is loaded into a register properly.

restore_full_context added checking to ensure the loaded PC is less
than 0x8100000