Fix hang on screen disconnect
usbdev_write would hang if the host screen session was closed,
persumably because the host was NAKing the device.

Now, if the write times out, it will abort the previous transaction and
try again.  As long as the NAKs continue, the transactions will continue
to be aborted.  The result is that everything can continue to run,
though aborted data is lost.