Buffer vfprintf characters before printing
When individual characters are being emitted, they are placed in a
holding buffer instead of through putc.  This holding buffer is flushed
with puts when it is full, at the end of the function, or before a
format string must call puts to emit its output.

This speeds up output on stm32f40x USB by ~25%.