Fix abandoned semaphores test.
In the event that a task failed to release a semaphore before ending,
and another task attempted to acquire that semaphore before kernel_task
released it on the deceased's behalf, an attempt would be made to switch
to the dead task, which obviously did not go well.

The solution to this is to check that the held_by task still exists
before switching to it.  Unfortunately, this means walking the task
list on every acquire failure, but I see no alternative.

One improvement to add is to immediately switch to the kernel task in
the event of this happening, however there currently is no simple way of
switching to a periodic task which may or may not be running or in the
task_list.