Resource overhaul, adding error checking every level.
All of the primary resource functions (read, write, close, etc) will
return success or failure values.  read/write return number of chars
written/read, while close returns zero on success.  This error checking
goes all the way down to the resource driver level, where all resource
functions return in the same manner as the parent function.

The one exception in the resource level read().  Since it returns the
char that was read, it is passed an int *error, which it should set to
zero on success, nonzero on failure.

The addition of error checking through the entire resource system makes
the entire system much more robust, and eliminates all panics in the
system.

We have have errors, now **PLEASE** actually check for them!