PX4: MS5611 barometer driver
Adds support for MS5611 barometer on PX4, which provides temperature and
pressure.  The values I am getting are sane, though they seem about
20mbar low today, based on the nearby weather stations.  The low
temperature corrections are untested, as it is far to warm today to test
without going into a freezer.

Also, there are sketchy calls to yield in here, as the driver has to
wait 10ms for each measurement, but we have no sleep currently.