Add usleep function for waiting
usleep uses the new system_ticks global, and waits for it to exceed a
certain value.  Note that the precision of usleep is limited to the
systick period, which defaults to 250us.

There is also overflow checking, as system_ticks may overflow in only 12
days @ 250us ticksAdd usleep function for waiting

usleep uses the new system_ticks global, and waits for it to exceed a
certain value.  Note that the precision of usleep is limited to the
systick period, which defaults to 250us.

There is also overflow checking, as system_ticks may overflow in only 12
days @ 250us ticks.