Add generic tracing mechanism
The TRACE(args...) macro expands to fprintf if tracing is enabled, and
nothing if it is disabled.  By default, ITM will be used for tracing, if
supported.  Otherwise, stderr will be used.  The ENABLE_TRACE Kconfig
option is used to control tracing.

Also, slightly reordered some of the SWD/ITM Kconfig options.  The
ALLOW_SWD option has been removed, as it was redundent.  HAVE_SWD
implicitly gives permission to use it.