Switch to fully recursive build system
Each folder now has its own Makefile, however the advantage is that all
of the Makefiles are very simple.

Each Makefile need only define SRCS or SRCS_y with the sources to be
built and DIRS or DIRS_y with the directories to be built.  Finally,
include "$(BASE)/tools/submake.mk" at the end of the Makefile.

This include will handle all building, linking, and recursing into other
directories.

Please note that source files *must* be in the same directory as the
Makefile, and directories to be built should be a single level down from
the current directory.