Fix abort periodic task removal
Aborted periodic tasks were not appropriately removed from the periodic
task list, causing them to hang the system at a later time.  This was
because the task end code attempted to remove the *standard* task list
node from the periodic task list node.  The node was not in the list,
but remove_task fails silently.