Add strncmp test, fix edge case
In the case where the (n+1)th character did not match, strncmp would
return not a match, even though the first n characters matched.  This is
because when breaking out of the while loop, the pointers had already
been updated to the next letter, but there was no check to return a match
if n == 0.