sched: Make scheduler more independent
Make the OS significantly less dependent on the internal structure of
the scheduler.  Now, curr_task is a task_t, which is a subset of the
entire task structure, which contains task-specific structures needed
outside of the scheduler.  Namely, per-task structures for semaphores
and resources.  Internally, the task_t can be mapped to the complete
task_ctrl structure using container_of.

Now, kernel/sched.h defines all externally available APIs to the
scheduler.  The entire structure and workings of the scheduler could be
swapped, as long as this API to maintained.  Currently, there is one
task comparison API.  In the future, additional task_t operations may be
added to the API.

Change-Id: I9ed0d337a474b324d6682424468f9751e55cc95a