sched: Add cooperative task scheduling
The new task_switch API function will allow a task to cooperatively
switch to another task.  The task being switched to must be runnable;
that is, it must still exist, and not be a periodic task between
periods.  task_switch will return an error if the task could not be
scheduled.

Change-Id: I6ad157d157256262714d44691cb6fb3236638433