lib: transform sprintf to scnprintf
Replace sprintf with scnprintf, which uses the updated buffer length API
for buffer streams.  This means that a length argument must be passed
along with the buffer to be written.

This function returns the number of bytes actually written to the
buffer, hence the name scnprintf instead of snprintf.  See
http://lwn.net/Articles/69419/ for more details.

Change-Id: I5072c7ea8a1686324ca6355635e00fdd99d81610