mm: bitfield allocator
This adds the "bitfield" allocator, which uses bitfields to track
free memory. It manages memory as a collection of blocks, where a
block is a collection of grains. A grain is the smallest amount of
memory that can be allocated. Blocks have a freemask, a bitfield to track
free grains. It has a configurable grain size, and divides memory
into these blocks, where each block covers 32 grains. The least
significant bit of the freemask corresponds to the first grain.

Change-Id: I00f127ea5cf56dc0e6f1858fb5ff88048dac6612