spi: remove spi address argument
Previously, spi_read and spi_write accepted an address argument, which
was written out before the data to be read/written.  This was in use
because many SPI devices protocol involves writing the address to be
read or written before doing so.

However, since this is simply a convention, rather than part of the SPI
protocol, it hinders development for devices which do not follow the
convention.  By removing the argument, we leave it up to individual
devices to write the address before performing a read or write.

Further complicating matters. unfortunately, is that spi_read and
spi_write automatically handle setting the chip select high and low,
meaning that CS will go high between writing an address and actually
reading.  In order to avoid this, the concept of extended transactions
is introduced.  Extended transactions take place over more than old call
to spi_read and/or spi_write, with CS held low the entire time.  It is
not safe to release control of the SPI port during an extended
transaction, as the CS of the device is still low.

Update device drivers on the SPI bus to use this new interface.

Remove SPI functions from the .kernel section.

Change-Id: I1c56af7b8acd5b932497f5451d450f947ed2f585