Prefix arch and chip includes
Prefix arch and chip specific includes with arch/ and arch/chip/,
respectively.

This is done by copying the relevant includes into an include folder in
the output directory, and using that as the root include directory.

Doing this provides distinct namespaces for arch and chip headers, which
will never collide with global headers.

Change-Id: I5cc01df3ac2460105d644b8ba6ec771a564139cf