Preprocess the linker script
Run the linker script through the C preprocessor to allow the use of
CONFIG_ macros in the linker script.  It is also possible to include
header files and use macros, however headers with C symbols in them may
cause the linker to complain.  It is the developer's responsibility to
ensure that the generated linker script is valid.

Change-Id: Ieb4e4550b1c9c8556bdd19683339bba7f618adc7