init: add "initializers"
More or less the same as initcalls on Linux (but simpler), initializers
provide a nice way for someone to initialize something at startup time.

There are early, core, and late initializers, all run in that order.

Early initializers CANNOT depend on any hardware or the heap to be initialized
when they run.

Change-Id: Iac2892f7247eb806779124be9c7e70010bde038c