stm32f4: obj SPI driver
Convert the existing STM32F4 SPI driver to the new obj-based SPI class.
Instead of using the SPI api, drivers should now keep a reference to the
SPI object of the bus they are on, and use its ops for interactions.

For now, all of the supported SPI ports are listed in an array, and
initialized at boot.

Change-Id: I2f9f9e2e66e3f4b65966880933b4f3448a82018d