dev: spi class
Provide a generic SPI class using the obj device framework.

Instead of providing a SPI API, use the obj framework to define a
SPI class, of which physical SPI implementations will be a member of.
Devices on a SPI bus will have a SPI object as their parent, and will
use that object's ops for all communications.

Remove the SPI API; all drivers should switch to the obj device framework.

Change-Id: Ic660b7fafdf40fee74aeef936f793b7545b4bc3a