dev: generic LED driver
The LED driver is a simple wrapper around the GPIO driver.  A few
board-specific files provide an array of the GPIOs used as LEDs, then
the led_ wrapper functions can be used to get one of the LEDs by index,
and enable, disable, and toggle it.

Since the LED GPIOs are now objs, the power LED cannot be enabled until
after the heap in initialized.

Change-Id: I1374488acb2d5f5084762cfdc4eb23871d2dc9f9