build: fixup parallel build
This fixes several known, and a few previously unknown issues with
building in parallel (-jN).

'make again' can now be run with parallel jobs.  To prevent the clean
rule from running in the middle of the build, explicit submake calls for
clean and all are made.

'make defconfig all' run in parallel would run into a similar issue of
the configuration changing in the middle of the build.  Work around this
by not allowing an 'all' build at the same time as a defconfig build.
Only the defconfig rule will run.

'make' run in parallel without a configuration would result in an error
attempting to preprocess the linker script.  The linker script rule is
now dependent upon the configuration.

The Kconfig generated files have been given explicit variables which can
be depended upon.  Additionally, the generated .conf file is included
instead of the .config file, which is not intended to be included in a
Makefile.  As a result, KCONFIG_MAKE_DEFS had to be given a dummy rule
to prevent Make from attempting to rebuild it during the include.  As
such, it cannot be depended on.  KCONFIG_HEADER must be depended upon
instead.  Both are generated by the same rule.

Generate the config headers when processing a defconfig, since .config
is not included.

Change-Id: Ic22ff469e190c346daf065ea0dc4a0866f3846d9