stm32f4: spi: switch to GPIO objects
Switch from directly setting the GPIOs to using the GPIO obj interface.
This brings the driver one step closer to having generic initialization.

Additionally, move the private data setup from init() to a 'constructor'
function.  This is not quite a full constructor, as it takes the special
SPI config struct, but it is yet another step towards a general setup.

Change-Id: I7b1930b929f898df6757a5088a2ea93e605a8af7