dev: driver registration
Provide device driver registration and look up.

Device drivers register themselves in an initializer using a
device_driver struct.  This struct includes such details as the device
name, class, a constructor to build the device object, and a probe to
ensure device existance.

The new device_get() is used to request a device, simply using the
device name.  The device driver list will be searched for a matching
device.  Once found, the device's class is searched for an existing
instance of the device.  If an instance cannot be found, a new instance
will be created using the device constructor.

Dependecies are handled as construction happens.  If on device is
dependent upon another, it will call device_get() in its construction,
which will go through the same construction process.

Some of the icky details:  There exists a race condition between looking
up an instance and construction.  If no instance is found, one must be
constructed.  However, an earlier entry may already be in the
constructor building said instance, it just hasn't been added to the
class yet.  In this case, a duplicate instance would be constructed,
which the driver may not be able to handle.

To solve this problem, driver semaphores are used to serialize the
construction of new instances of that device.  Individual driver
semaphores are perfered over a single global lock, since concurrent
construction of different devices is fine, it is simply duplicate
devices that are an issue.  In addition, some devices may take an
exessive amount of time to initialize, and forcing all device
initialization to wait on them is highly undesirable.

The primary caveat of this approach is that a device could not depend on
another device with a shared driver semaphore, as that would require a
reentrant mutex, since the second device_get() call would attempt to
acquire the same semaphore.  This should not be a large issue, as
different devices should not need to share a driver semaphore, and
devices that depend upon other instances of themselves should be able to
handle this dependency internally to the driver.

Change-Id: I33c717daa0e1d8b9324932b95cc37db3a0dbb829