kernel: Abstract arch-specific portions of sched and semaphores
Move all arch-specific components of the sched and semaphores to the
arch/ folder.  This meant providing exported definitions for many of the
sched functions, which are available in kernel/sched_internals.h.  The
symbols in that header should only be used by kernel/sched/, but must be
implemented by the architecture.

The semaphore try_lock has been replaced with the GCC builtin
__sync_bool_compare_and_swap().  In the future, this should be a
standard F4OS atomic operation.

Change-Id: I24df70dcc4141c74c5e59da6f20cb0f5b8c7c7de