arch: Initial port to ARMv7-A and TI AM335X
Currently, it must be manually loaded into memory.  It does not support
task switching, and thus ends execution with a "Task switching ended"
panic, but it does successfully run to completion.

Includes a barebones device tree source for the AM335X.  Since the Linux
dts is GPL'd, the current plan is to write our own DTS, using the same
bindings as the Linux kernel, to maintain compatibility.

For most atomic operations, use the __atomic_* GCC builtins.  These are
guaranteed to generate atomic accesses with the specified memory model.
For now, the most conservative memory model is selected, ensuring full
data synchronization.

Change-Id: Ie002c31ab4d12435bbaa4a0c6f9734289e4293a1