am335x: add initial IRQ and FIQ interrupt handlers
These handlers look up the given the source of the interrupt, and call
the handler out of a table of function pointers.

A standard set of functions is provided to enable, disable, register,
set priority, etc of interrupts, based on the interrupts and
interrupt-parent FDT properties.

The appropriate interrupt controller is looked up in interrupt-parent,
but currently only the primary system interrupt controller is supported.
This is the controller referenced by interrupt-parent in the root node.

Change-Id: I193f30244b8edc5dadea6abb7107b75dacdc3145