dev: add raw memory helper macros
Provide raw_mem_* macros to aid in safely reading and writing raw memory
locations.  These macros include memory barriers to ensure that accesses
are not reordered.

The this time, only a compiler memory barrier is used, not a CPU memory
barrier.

The memory barrier macros are also provided by this header.

Change-Id: I9e49dfa15c4467f2316119c0618e02c460f1dfd5