am335x: system tick timer
Enable and use the DMTimer 1ms for use as the system tick timer.
For now, simply the first DMTimer 1ms found in the device tree
will be used as the system timer.

Use the main oscillator as the source, computing the number of counts
necessary for the desired system tick frequency.

Set the load register to count that many ticks up to 1<<32 - 1, then
trigger the overflow interrupt.

In the system tick handler, perform the standard tick routines.

Change-Id: I5dd5988917fdbad9eb859e7672257ccdc3517119