stm32f40x: obj i2c driver
Convert the STM32F4 I2C to the new obj-based I2C driver class.  The
internal layout is similar to the STM32F4 SPI driver.

This made it very easy to add I2C3 support, as it is just an additional
array entry.

Currently, all I2C ports are configured the same, and limited to the
hard-coded GPIOs.

Change-Id: I979888852f8f57d0374d589c1fa90ce7770a2a63