obj: remove dtor field from class
Class objects have a dtor field which is completely unused.  Some class
users set to the destructor for objs in the class, some neglect it
completely.  Regardless, it is never accessed.

Since it is unnecessary and unused, remove it.  Should some dire need
arise in the future, it may be added back, but for now, it has no
defined use, and it is best to avoid having users depend on it in
unconventional ways.

Change-Id: Id32eae4388cb449e13dad7ebf4b45f9a6e1c027b