dev: barometer class
Add a generic barometer device class.  This driver is similar to the
accelerometer and magnetometer classes, with only a few differences.

Since many, but not all, barometers provide temperature in addition to
pressure, include it in the baro_data struct, but provide a method to
determine if a barometer supports temperature.

Additionally, since barometers almost universally require lots of
internal conversions and math, do not provide raw data, as it has no
universal format.

Change-Id: I2a2212eded9ec8f315b51ba20e3cf1c4bbe73237