dev: release driver semaphore when instance found
When an instance of a device was found in its class, device_get()
immediately returned, leaving the driver semaphore held by the current
task, making it impossible for more instances to be retrieved.

Additionaly, properly handle non-existent drivers

Previously, driver would never be NULL after looping through the
drivers, so the last driver in the list would be used instead of
returning NULL.

Fix that to properly handle non-existent drivers.

Change-Id: Icb729b21a71bc7416003ad8b568a7c8ea3334a6b