kernel: check for acquires of held tasks
When deadlock checking, explicitly check for attempts by a task to
acquire a semaphore they already hold.  In the best case, it will cause
an infinite loop.  In the worst case, it may crash the system if it is
unable to do a task switch to the currently running task.

Change-Id: I61ecd628af6837d5e8714bc15b017f690035abf8