periph: MPU6000 accelerometer and gyroscope
Convert the PX4 MPU6000 driver to generic obj accel and gyro drivers.

Since the MPU6000 is both an accelerometer and gyroscope, this
conversion was a bit trickier than previous.

A new MPU6000 class is added, which provides generic functionality for
accessing the MPU6000 registers and enabling and disabling the
accelerometer and gyroscope.  A SPI instance of this class is provided,
configured with the board config files.  In the future, an I2C instance
could provide transparent support for the I2C MPU6000.

Depending on the MPU6000 class instance are the MPU6000 accel and gyro
drivers.  These are standard accelerometer and gyro class instances,
except instead of depending on a SPI or I2C bus, like most drivers, they
depend on the MPU6000 defined above.  They use the MPU6000 interface to
access the registers they need for configuration and data retrieval.

Change-Id: I44639415b28a7bc8ceb0f73e0e05a2867c63faaa