shell: automatically detect accelerometer driver
Make the accel program automatically detect the name of the
accelerometer driver.  If there are none, or multiple drivers, present
an error to the user, listing the drivers if there are multiple.

This may get transformed into a generic device shell program to avoid
copying this functionality to the gyro, mag, and baro programs.

Change-Id: Ifa6e96aeac5ee7c9f8841daa7e616fa96aade26a