build: remove kconfig-frontends from source tree
Instead of distributing a snapshot of kconfig-frontends in the F4OS
source tree, make it an external dependency.

Since most Linux distributions do not provide a package for kconfig,
include a script to automatically download and install the tools.  This
means that the build flow is similar to before, except that new clones
must download kconfig-frontends, in addition to configuring and
building.

The build system will first look for the conf tools in the PATH.  If
they are not found there, it will fallback to the custom build.

Change-Id: I11d91456406745763b23f472b9c07644cd2ec6b0