periph: InvenSense ITG-3200 gyroscope driver
Convert the old "9DOF gyro" driver to a proper obj driver for the
InvenSense ITG-3200 gyroscope.

This driver is configured with the board config structs, just like the
others.

Since this peripheral is not permanently connected to the Discovery
board, remove support from the Discovery defconfig.  It can be enabled
in the Gyroscope submenu in Driver config.

Since this driver is in the gyro class, it does not need its own shell
program.  It will work with the standard "gyro" program.

This is the last external stm32f40x peripheral to be converted!  As
such, arch/armv7m/chip/stm32f40x/dev/periph/ is removed entirely.

Change-Id: I7ce09bdcebf7e859c20046289a4969996739c6f2