lm4f120h5: gpio driver
Add an obj-base GPIO driver for the LM4.  This generic GPIO driver
replaces the old LED-only driver with a generic GPIO driver.

It adds support for the standard LED framework on the LM4, allowing
the LM4 to build successfully with the obj framework.

An unconditional board file is used to define the default LEDs.

Change-Id: I5836d7de49991b097851cf7045e5e42f2dc34d25