dev: add missing raw_mem_read parentheses
raw_mem_read was missing parentheses around the use of the __address
arguement.  As a result, operations done in the arguement, may not
behave as expected.  For example:

raw_mem_read(&base + 0x10);

This should do pointer arithmetic on base, and read that address.
Instead, it read base, then added 0x10 to the result.  Not the expected
behavior.

Change-Id: I6e08e5df918e8e3265fd1f55d2846b140d496052