sched: use standard yield to begin task switching
Rather than implementing restore_full_task() to emulate a full processor
context restore when starting the first task, simply make a call to
task_switch() to yield to the scheduler.

The task switch service call handler detects that this is the first
call, and sets task switching enabled.

On ARMv7-M, this initial service call is made while the MSP is active,
rather than the PSP.  The SVC handler detects this, and passes the
correct register set to svc_handler(), and fixes up the LR to allow
exception return to the PSP, once a user task is set.

Change-Id: Ifc4c1f8eecc512ae34d68108099acbcf6e3e6aa0