armv7m: sched: ensure stack alignment on context creation
AAPCS (Procedure Call Standard for the ARM Architecture) requires that
the SP be double-word (8-byte) aligned at a public interface.  Entry to
the first task function is a public interface, so the SP must be
aligned.

When creating task context, force the stack pointer to an 8-byte aligned
address.  The actual context creation instructions will add an even
number of words to the stack, ensuring that the SP is still aligned at
the top of the context.  The FPU context will add an addition empty word
to ensure alignment.

Change-Id: I85a4fcf0dd7952f3dde857c806e12bdbcee21f0b