config: conditionally build with FPU/short double support
Use CONFIG_HAVE_FPU and (the new) CONFIG_SHORT_DOUBLE to determine
whether or not to build with those options, for Cortex-M4F.  If either
is disable, libgcc in linked in to provide floating point arithmetic
routines.

Additionally, drop the double promotion warnings.  Some promotions (such
as in variadic functions), are unavoidable when building without short
doubles.

Change-Id: I05d13cc470dd9a8876907913bf9f07c59956aea8