stm32f40x: i2c: split initialization
Create a new initialize function that handles the actual initialization,
and make the init method simply grab the I2C lock and then call this
function.

This way, other uses of the initialization, such as in reset, or when
the I2C port is not ready do not need to release the lock before calling
the function.

Change-Id: I52e2dd9015475c75c56b38f8d91cccd6cf0f51f7
(cherry picked from commit 303097729a02484dcb21e4cd20ea957cb3b61eac)