dev: AMS AS5048B driver
Add a driver for the AMS AS5048B I2C magnetic rotary encoder.  This
rotary encoder fulfills the standard rotary encoder interface.

The calibration used is software calibration, only valid for the life of
the object.

The standard board config struct is used, however instead of placing
config structs in each board file, define a weak, invalid, one in the
driver itself, which can be overridden by a board config.  This not only
keeps the board files clean, but saves data when the driver is not built
in.

Change-Id: I307d193526efc33b526573e826faefe862ecb73e