stm32f40x: perfcounter: avoid overflow race condition
Since TIM1_CNT and TIM2_CNT are not read atomically, it is possible for
TIM1_CNT to overflow after TIM2_CNT has been read, resulting in an
incorrect complete timer value.

In order to avoid this race condition, ensure that the TIM2_CNT has not
changed after reading TIM1_CNT.  If it has, TIM1_CNT has overflowed and
the read should be restarted.

This depends on the assumption that TIM1 will not always overflow in the
time it takes to go through the loop, otherwise it will be impossible to
break out.  TIM1 is only overflowing every 390us with a 168MHz clock, so
it should be fine.

A preemptive task switch may occur between reading the upper and lower
registers.  Such an event makes it more likely that the loop will fail
upon return to the task, but does not reduce the safety of this loop.

Change-Id: I0457ca7991f3f5e99365f98b482aa05c03d9f4b6
(cherry picked from commit d3f7008ae62e15718c36150bbd2166be9d7390a2)