stm32f40x: hardware pwm driver
Add driver for hardware PWM.  This driver uses timers
1,3,4,8,9,10,11,12,13,14 in the stm32f40x to provide PWM outputs for a
large number of GPIOs.

Each GPIO is mapped to the set of timer channels that can be used to
drive the output.  When creating a PWM object, these channels are
searched for a free channel that is either on an unused timer, or a
timer already configured to the desired duty cycle, as duty cycle is
configured for the entire timer.

The driver is rife with switch case statements used to generalize it for
all of the timers, but the core logic is not too complicated.  The timer
is configured in a simple upcounting mode with the desired duty cycle,
then the output compare units are set up with the desired pulse widths.

Change-Id: If762af0447cab6e329271f08866f2882dd139d55
(cherry picked from commit 961f9a93f4b53a61d15f216c976f8e87edbbbbf1)