kernel: build: include function symbol table
Build a symbol table, containing function names and addresses, into the
OS image itself.

A Python script parses output of `objdump -t` and dumps the names and
addresses of all functions in the symbol table.  Sorts by address, and
places addresses and names in an array.  This way, the array can be
searched by address for the nearest matching symbol.

Building this symbol table into the OS requires two passes.  One to
create a symbol table of the correct length, and a second to fixup the
addresses changed by the change in symbol table size from the first
pass.

Building the symbol table into the OS is controlled by the
CONFIG_BUILD_SYMBOL_TABLE config option.  When disabled, only the
default empty table is included.

A function is included to lookup a function name at runtime.  The
address of an instruction in the function is given, and the symbol name
just above the instruction is returned.

Change-Id: I8cd65cad2e45dec893cde6aab9eb7175e0a6e54f