class: instantiate() does not return allocated address
instantiate() allocates a class instance by allocating space for its
container type, and returns a pointer to the struct obj within the
container.  The obj is returned because it is the universally know type
shared between classes.

However, when freeing a class instance, care must be taken.  It is the
*container* that needs to be freed, not the obj returned from
instantiate().  Thus, if you only have the obj, get_container() needs to
be used to find the correct container address for freeing.

Add instantiate() documentation to explain this.

Change-Id: I0e30b68759301928b0b6012b37fda3d72372a6f2