config: move generic stm32f4 contents to dedicated file
Since the chip peripherals are common to all stm32f4 chips, put them in
stm32f4.dtsi, which is included by the discovery board and px4 device
trees.

Rename the px4 device tree to stm32f4_px4.dts, to indicate that it is an
stm32f4.  Also, it is more specifically the px4fmu.

Change-Id: I882b55dc7d780372964739caeb9fd548378eadff