dev: ms5611: convert to FDT
Convert the MS5611 barometer driver from a board_config.h-based
configuration to an FDT-based configuration.

There is no unique identifying register in the chip, so simply perform a
dummy write to the device, looking for an error from the I2C peripheral,
indicating that there is no device at that address.

Change-Id: I5cf5c497a0d1730a36caeed7b87f7f01b2daaba6