dev: mpu6000: convert to FDT
Convert the MPU6000 gyro/accelerometer from a board_config.h-based
configuration to an FDT-based configuration.

The MPU6000 node is a subnode of the appropriate bus.  The accelerometer
and gyro components are subnodes of the MPU6000.  While this isn't
strictly necessary from a hardware perspective, as all MPU6000s have an
accelerometer and gyro, it provides automatic enumeration of the
drivers, like all other devices in the device tree.  The MPU6000 accel
and gyro drivers can then connect to the correct MPU6000 driver by
simply looking up their parent.

For MPU6000 probing, verify the contents of the WHOAMI register.  For
accel and gyro probing, simply verify that the parent MPU6000 is
accessible.

Change-Id: I0bad56f18d91bb76615fd44fabfac594f8bd45ad