class: strdup name on instantiation
To avoid issues with use-after-free or storing a stack address with the
class instance obj name, unconditionally strdup() the name argument to
instantiate().  Thus, we are guaranteed to have ownership of the name.

Since there is now more than one thing to free when destorying a class
instance, provide class_deinstantiate() for destroying these objects.

As a small addition, provide class_unexport_member(), which performs the
opposite of class_export_member(), and should typically be called before
calling class_deinstantiate().

Change-Id: If630442618db658c3fa106e0d53dbbf71c877f08