am335x: add prcm clocks driver
Add a clocks driver for the AM335x PRCM module, which controls clock
gating to modules.  Additionally, the set_param() function can be used
to select clock sources for several module clocks.  It simply assigns
the parameter to the register.

In the device tree source, describe the hierarchal L3 and L4 bus
interconnects.  In the PRCM node, each of its register banks is
described by a subnode.

Change-Id: I4b56d26a89c0819b7ed42e80ea6bd69b3815f414