armv7a: initialize all priviledged stack pointers, panic exceptions
At startup, set the SP for SVC, IRQ, FIQ, ABR, and UND modes to
configured values.  This way, the SP will already be set up when those
exceptions are taken.

When an abort occurs, branch to a handler, which will currently panic
the system.

In addition, for the AM335x, make more efficient use of the 63K internal
SRAM + 64K of L3 OCMC ram.  Allocate 32K for user heap, 16K for kernel
heap, and 1K of stack space for each priviledged mode, except SVC, which
gets 12K.

Change-Id: I3e6c821352fd0cda97ea8a1224340c3f301e8903