armv7m: enable floating point extension at startup
Previously, the automatic floating point state preservation bit was set
in the floating point context control register.  This bit enables
floating point state preservation (automatic FP register saving on
exception entry) *only once the first FP instruction is executed*.

As a result, before the first FP instruction was executed, a task's
context would be saved without FP state.  If it was restored after an FP
instruction was executed, the core would "restore" the FP state, thus
corrupting the SP by "restoring" crucial parts of the task's stack.

Alternatively, set the FPCA in the CONTROL register at chip startup.
This enables FP state preservation from the start, and is the the bit
set by the core when the first FP instruction is executed when the ASPEN
bit is set.

Change-Id: I1e3456c4a150c52183e1a220fa74602f1ca88631